+++ chain/client/src/stateless_validation/validate.rs
@@ -1,11 +1,9 @@
 use super::partial_witness::witness_part_length;
-use crate::metrics;
 use itertools::Itertools;
+use near_chain::signature_verification::resolve_and_verify_anchored_producer;
 use near_chain::types::Tip;
 use near_chain_primitives::Error;
-use near_epoch_manager::{CHUNK_GRANDPARENT_ANCHOR_HEIGHT_OFFSET, EpochManagerAdapter};
-use near_primitives::errors::EpochError;
-use near_primitives::hash::CryptoHash;
+use near_epoch_manager::EpochManagerAdapter;
 use near_primitives::stateless_validation::ChunkProductionKey;
 use near_primitives::stateless_validation::chunk_endorsement::ChunkEndorsement;
 use near_primitives::stateless_validation::contract_distribution::{
@@ -14,15 +12,21 @@ use near_primitives::stateless_validation::contract_distribution::{
 use near_primitives::stateless_validation::partial_witness::{
     MAX_COMPRESSED_STATE_WITNESS_SIZE, VersionedPartialEncodedStateWitness,
 };
-use near_primitives::types::validator_stake::ValidatorStake;
-use near_primitives::types::{AccountId, BlockHeight, BlockHeightDelta, EpochId, ShardId};
+use near_primitives::types::{AccountId, BlockHeightDelta};
 use near_primitives::validator_signer::ValidatorSigner;
 use near_primitives::version::ProtocolFeature;
-use near_store::{DBCol, FINAL_HEAD_KEY, HEAD_KEY, Store, get_genesis_height};
+use near_store::{DBCol, FINAL_HEAD_KEY, HEAD_KEY, Store};
 use strum::IntoStaticStr;
 
 /// This is taken to be the same value as near_chunks::chunk_cache::MAX_HEIGHTS_AHEAD, and we
-/// reject partial witnesses with height more than this value above the height of our current HEAD
+/// reject partial witnesses with height more than this value above the height of our current HEAD.
+///
+/// Note: for V2 witness parts (which carry the grandparent anchor), arrival-time producer
+/// resolution requires the anchor to be processed, so a witness more than
+/// CHUNK_GRANDPARENT_ANCHOR_HEIGHT_OFFSET (2) heights ahead of the head has an unprocessed
+/// anchor and is dropped as "node behind" — making the *effective* horizon 2 for those,
+/// even though this constant stays 5. Keeping it at 5 leaves V1 / pre-feature behavior
+/// unchanged. See docs/design/chunk_ancestor_signature_verification.md.
 const MAX_HEIGHTS_AHEAD: BlockHeightDelta = 5;
 
 /// This enum represents whether a particular chunk is relevant in the context of validating
@@ -137,143 +141,6 @@ pub fn validate_partial_encoded_state_witness(
     }
 
     Ok(ChunkRelevance::Relevant)
-}
-
-/// Cross-check the signed (epoch, height) against the signed anchor before trusting the
-/// producer resolved from it. The anchored lookup is keyed by anchor and shard only, not
-/// height, so without this a producer could sign under any height and we would store or
-/// forward under a forged key.
-///
-/// Parent known: must match the parent exactly. Parent absent, real anchor: height must be
-/// anchor + 2. Parent absent, default anchor: only valid at genesis or genesis + 1.
-/// `msg_label` names the caller in the error.
-fn verify_anchored_chunk_key(
-    epoch_manager: &dyn EpochManagerAdapter,
-    epoch_id: &EpochId,
-    height_created: BlockHeight,
-    prev_block_hash: &CryptoHash,
-    prev_prev_block_hash: &CryptoHash,
-    store: &Store,
-    msg_label: &str,
-) -> Result<(), Error> {
-    match epoch_manager.get_block_info(prev_block_hash) {
-        Ok(parent_info) => {
-            let expected_epoch_id = epoch_manager.get_epoch_id_from_prev_block(prev_block_hash)?;
-            if parent_info.prev_hash() != prev_prev_block_hash
-                || parent_info.height() + 1 != height_created
-                || &expected_epoch_id != epoch_id
-            {
-                return Err(Error::InvalidPartialChunkStateWitness(format!(
-                    "V2 {msg_label} chunk key mismatch: signed (epoch_id={:?}, height={}, \
-                     prev_prev={:?}) does not match prev_block_hash-implied \
-                     (epoch_id={:?}, height={}, prev_prev={:?})",
-                    epoch_id,
-                    height_created,
-                    prev_prev_block_hash,
-                    expected_epoch_id,
-                    parent_info.height() + 1,
-                    parent_info.prev_hash(),
-                )));
-            }
-        }
-        Err(EpochError::MissingBlock(_)) => {
-            if prev_prev_block_hash != &CryptoHash::default() {
-                // Parent not here yet, so only the anchor is known. Requiring height ==
-                // anchor + 2 ties one anchor to one chunk key per shard, so a producer
-                // cannot reuse it across many cache slots. (Cross-epoch anchors fall back
-                // to the resolver.) A skipped slot with a missing parent is dropped here;
-                // that is fine, best-effort.
-                let anchor_height = epoch_manager.get_block_info(prev_prev_block_hash)?.height();
-                let expected_height = anchor_height + CHUNK_GRANDPARENT_ANCHOR_HEIGHT_OFFSET;
-                if height_created != expected_height {
-                    return Err(Error::InvalidPartialChunkStateWitness(format!(
-                        "V2 {msg_label} height {height_created} does not match \
-                         anchor-implied height {expected_height}"
-                    )));
-                }
-            } else {
-                // Default (genesis) anchor with no parent: nothing pins the height. A real
-                // default anchor only happens at genesis or genesis + 1, so reject higher
-                // to avoid an any-height hole.
-                let genesis_height = get_genesis_height(store)
-                    .ok_or_else(|| Error::Other("genesis height not found".to_owned()))?;
-                if height_created > genesis_height + 1 {
-                    return Err(Error::InvalidPartialChunkStateWitness(format!(
-                        "V2 {msg_label} with default anchor at height {height_created} \
-                         above genesis + 1 ({})",
-                        genesis_height + 1
-                    )));
-                }
-            }
-        }
-        Err(err) => return Err(err.into()),
-    }
-    Ok(())
-}
-
-/// Look up the producer from the anchor and record the lookup-outcome metric (hit, anchor
-/// block missing, or `DBCol::ChunkProducers` row missing). Error returned unchanged;
-/// callers turn the node-behind cases (`MissingBlock`, `ChunkProducerNotInDB`) into a
-/// quiet drop.
-fn resolve_anchored_producer(
-    epoch_manager: &dyn EpochManagerAdapter,
-    prev_prev_block_hash: &CryptoHash,
-    epoch_id: &EpochId,
-    height_created: BlockHeight,
-    shard_id: ShardId,
-    message_type: &str,
-) -> Result<ValidatorStake, EpochError> {
-    let result = epoch_manager.get_chunk_producer_info_anchored(
-        Some(prev_prev_block_hash),
-        epoch_id,
-        height_created,
-        shard_id,
-    );
-    let label = match &result {
-        Ok(_) => "hit",
-        // Anchor block not processed yet: this node is two or more blocks behind.
-        Err(EpochError::MissingBlock(_)) => "miss_anchor_block",
-        // Anchor is processed but has no `DBCol::ChunkProducers` row. Should be ~0
-        // normally; if it persists, something that writes that row has a bug.
-        Err(EpochError::ChunkProducerNotInDB(_, _)) => "miss_db_entry",
-        Err(_) => "error",
-    };
-    metrics::ANCHORED_CHUNK_PRODUCER_LOOKUP_TOTAL
-        .with_label_values(&[shard_id.to_string().as_str(), message_type, label])
-        .inc();
-    result
-}
-
-/// Resolve a V2 message's producer from the grandparent anchor, then cross-check the signed
-/// chunk key against the anchor. Shared by the witness, contract-accesses and contract-deploys
-/// V2 paths; `msg_label` names the caller in the metric and errors. Resolve runs before the
-/// cross-check so the lookup metric is recorded even when the key check later rejects.
-fn resolve_and_verify_anchored_producer(
-    epoch_manager: &dyn EpochManagerAdapter,
-    key: &ChunkProductionKey,
-    prev_block_hash: &CryptoHash,
-    prev_prev_block_hash: &CryptoHash,
-    store: &Store,
-    msg_label: &str,
-) -> Result<ValidatorStake, Error> {
-    let producer = resolve_anchored_producer(
-        epoch_manager,
-        prev_prev_block_hash,
-        &key.epoch_id,
-        key.height_created,
-        key.shard_id,
-        msg_label,
-    )?;
-    verify_anchored_chunk_key(
-        epoch_manager,
-        &key.epoch_id,
-        key.height_created,
-        prev_block_hash,
-        prev_prev_block_hash,
-        store,
-        msg_label,
-    )?;
-    Ok(producer)
 }
 
 pub fn validate_partial_encoded_contract_deploys(
